Remove Streamlet configurations creation
