+++ core/cloudflow-operator/src/main/scala/cloudflow/operator/Main.scala
@@ -58,7 +58,6 @@ object Main extends {
       )
       Operator.handleAppEvents(client, runners, ctx.podName, ctx.podNamespace)
       Operator.handleConfigurationUpdates(client, runners, ctx.podName)
-      Operator.handleConfigurationInput(client, ctx.podNamespace)
       Operator.handleStatusUpdates(client, runners)
     } catch {
       case t: Throwable =>
